+++ bapsf_motion/gui/configure/controllers.py
@@ -2,12 +2,19 @@
 
 import logging
 import numpy as np
+import os
 import re
 import warnings
 
+# ensure joystick events are monitored when the pygame window
+# is not in focus ... this needs to be done before importing pygame
+os.environ["SDL_JOYSTICK_ALLOW_BACKGROUND_EVENTS"] = "1"
+
+import pygame  # noqa
+
 from abc import abstractmethod
-from PySide6.QtCore import Signal, QTimer, Qt, Slot
-from PySide6.QtGui import QDoubleValidator
+from PySide6.QtCore import Signal, QTimer, Qt, Slot, QThreadPool
+from PySide6.QtGui import QDoubleValidator, QFont
 from PySide6.QtWidgets import (
     QWidget,
     QLabel,
@@ -17,12 +24,14 @@ from PySide6.QtWidgets import (
     QGridLayout,
     QLayout,
     QSizePolicy,
+    QComboBox,
 )
 from typing import List
 
 from bapsf_motion.actors import MotionGroup, Drive, Axis, Motor
 from bapsf_motion.gui.configure.helpers import gui_logger
 from bapsf_motion.gui.configure.message_boxes import LostConnectionMessageBox
+from bapsf_motion.gui.configure.pygame_ import PyGameJoystickRunner
 from bapsf_motion.gui.icons import icon_name_dict
 from bapsf_motion.gui.widgets import (
     EnableIndicator,
@@ -31,6 +40,7 @@ from bapsf_motion.gui.widgets import (
     StyleButton,
     ZeroButton,
     HLinePlain,
+    LED,
 )
 from bapsf_motion.utils import units as u
 
@@ -1154,3 +1164,309 @@ class DriveDesktopController(DriveBaseController):
         self.hold_current_btn.setEnabled(True)
 
         super().enable_motion_buttons()
+
+
+class DriveGameController(DriveBaseController):
+    def __init__(self, parent=None):
+        super().__init__(axis_display_mode="readonly", parent=parent)
+
+    def _connect_signals(self):
+        super()._connect_signals()
+
+        self.refresh_controller_list_btn.clicked.connect(self.refresh_controller_combo)
+        self.connect_btn.clicked.connect(self.connect_controller)
+        self.controller_combo_widget.currentIndexChanged.connect(
+            self.disconnect_controller
+        )
+
+    def _initialize_widgets(self):
+        self._pygame_joystick_runner = None  # type: PyGameJoystickRunner | None
+        self._thread_pool = QThreadPool(parent=self)
+
+        # BUTTON WIDGETS
+        _btn = StyleButton("Refresh List", parent=self)
+        _btn.setFixedHeight(32)
+        _font = _btn.font()
+        _font.setPointSize(12)
+        _btn.setFont(_font)
+        self.refresh_controller_list_btn = _btn
+
+        _btn = StyleButton("Connect", parent=self)
+        _btn.setFixedHeight(32)
+        _btn.setFont(_font)
+        _btn.setFixedWidth(100)
+        self.connect_btn = _btn
+
+        # TEXT/ICON WIDGETS
+        _led = LED(parent=self)
+        _led.set_fixed_height(24)
+        self.connected_led = _led
+
+        # ADVANCED WIDGETS
+        _combo = QComboBox(parent=self)
+        _combo.setEditable(True)
+        _combo.lineEdit().setReadOnly(True)
+        _combo.lineEdit().setAlignment(
+            Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter
+        )
+        _combo.setFixedHeight(32)
+        _combo.setFont(_font)
+        self.controller_combo_widget = _combo
+
+    def _define_layout(self) -> QLayout:
+        self.refresh_controller_combo()
+
+        connect_layout = QHBoxLayout()
+        connect_layout.setContentsMargins(0, 0, 0, 0)
+        connect_layout.addStretch(1)
+        connect_layout.addWidget(self.connect_btn)
+        connect_layout.addWidget(self.connected_led)
+        connect_layout.addStretch(1)
+
+        _label_font = QFont()
+        _label_font.setPointSize(12)
+        _left_stick = QLabel("Left Stick :", parent=self)
+        _left_stick.setFont(_label_font)
+        _right_stick = QLabel("Right Stick :", parent=self)
+        _right_stick.setFont(_label_font)
+        _dpad_vert_stick = QLabel("DPad Up/Down :", parent=self)
+        _dpad_vert_stick.setFont(_label_font)
+        _dpad_horz_stick = QLabel("DPad Left/Right :", parent=self)
+        _dpad_horz_stick.setFont(_label_font)
+        _ab = QLabel("A / B :", parent=self)
+        _ab.setFont(_label_font)
+        _y = QLabel("Y :", parent=self)
+        _y.setFont(_label_font)
+        _move_y = QLabel("Move Y", parent=self)
+        _move_y.setFont(_label_font)
+        _move_x = QLabel("Move X", parent=self)
+        _move_x.setFont(_label_font)
+        _fine_y = QLabel("Fine Y", parent=self)
+        _fine_y.setFont(_label_font)
+        _fine_x = QLabel("Fine X", parent=self)
+        _fine_x.setFont(_label_font)
+        _stop = QLabel("STOP", parent=self)
+        _stop.setFont(_label_font)
+        _zero = QLabel("Zero", parent=self)
+        _zero.setFont(_label_font)
+
+        btn_label_layout = QGridLayout()
+        btn_label_layout.setContentsMargins(0, 0, 0, 0)
+        btn_label_layout.setColumnMinimumWidth(1, 8)
+        btn_label_layout.addWidget(
+            _left_stick, 0, 0, alignment=Qt.AlignmentFlag.AlignRight
+        )
+        btn_label_layout.addWidget(
+            _right_stick, 1, 0, alignment=Qt.AlignmentFlag.AlignRight
+        )
+        btn_label_layout.addWidget(
+            _dpad_vert_stick, 2, 0, alignment=Qt.AlignmentFlag.AlignRight
+        )
+        btn_label_layout.addWidget(
+            _dpad_horz_stick, 3, 0, alignment=Qt.AlignmentFlag.AlignRight
+        )
+        btn_label_layout.addWidget(_ab, 4, 0, alignment=Qt.AlignmentFlag.AlignRight)
+        btn_label_layout.addWidget(_y, 5, 0, alignment=Qt.AlignmentFlag.AlignRight)
+
+        btn_label_layout.addWidget(_move_y, 0, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+        btn_label_layout.addWidget(_move_x, 1, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+        btn_label_layout.addWidget(_fine_y, 2, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+        btn_label_layout.addWidget(_fine_x, 3, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+        btn_label_layout.addWidget(_stop, 4, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+        btn_label_layout.addWidget(_zero, 5, 2, alignment=Qt.AlignmentFlag.AlignLeft)
+
+        sub_layout_1 = QVBoxLayout()
+        sub_layout_1.setContentsMargins(0, 0, 0, 0)
+        sub_layout_1.addSpacing(16)
+        sub_layout_1.addWidget(self.refresh_controller_list_btn)
+        sub_layout_1.addWidget(self.controller_combo_widget)
+        sub_layout_1.addLayout(connect_layout)
+        sub_layout_1.addSpacing(24)
+        sub_layout_1.addLayout(btn_label_layout)
+        sub_layout_1.addStretch(1)
+
+        sub_widget_1 = QWidget(parent=self)
+        sub_widget_1.setLayout(sub_layout_1)
+        sub_widget_1.setMaximumWidth(200)
+        sub_widget_1.setMinimumWidth(100)
+        sub_widget_1.setSizePolicy(
+            QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Preferred
+        )
+
+        layout = QHBoxLayout()
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.addWidget(sub_widget_1)
+        layout.addSpacing(2)
+        for acw in self._axis_control_widgets:
+            layout.addWidget(acw)
+            layout.addSpacing(2)
+        layout.addStretch()
+
+        return layout
+
+    @property
+    def available_controllers(self) -> List[pygame.joystick.JoystickType]:
+        _joystick = pygame.joystick
+
+        if not _joystick.get_init():
+            _joystick.init()
+
+        return [_joystick.Joystick(_id) for _id in range(_joystick.get_count())]
+
+    @property
+    def joystick(self) -> pygame.joystick.JoystickType | None:
+        js_name = self.controller_combo_widget.currentText()
+        self.logger.info(f"Selected joystick: {js_name} - {self.available_controllers}")
+        js = None
+        for _js in self.available_controllers:
+            if _js.get_name() == js_name:
+                js = _js
+                break
+
+        return js
+
+    @Slot()
+    def refresh_controller_combo(self):
+        self.disconnect_controller()
+
+        current_controller_name = self.controller_combo_widget.currentText()
+
+        self.controller_combo_widget.clear()
+
+        controller_names = [
+            controller.get_name() for controller in self.available_controllers
+        ]
+        controller_names.append("")
+        self.controller_combo_widget.addItems(controller_names)
+
+        if current_controller_name != "" and current_controller_name in controller_names:
+            self.controller_combo_widget.setCurrentText(current_controller_name)
+            self.connect_controller()
+        else:
+            self.controller_combo_widget.setCurrentText("")
+
+    @Slot()
+    def connect_controller(self):
+        self.logger.info("Connecting controller.")
+
+        if isinstance(self._pygame_joystick_runner, PyGameJoystickRunner):
+            self.disconnect_controller()
+
+        self._pygame_joystick_runner = PyGameJoystickRunner(self.joystick)
+
+        self._pygame_joystick_runner.signals.joystickConnected.connect(
+            self._update_connect_led
+        )
+        self._pygame_joystick_runner.signals.axisMoved.connect(self._handle_axis_move)
+        self._pygame_joystick_runner.signals.buttonPressed.connect(
+            self._handle_button_press
+        )
+        self._pygame_joystick_runner.signals.hatPressed.connect(self._handle_hat_press)
+        self._pygame_joystick_runner.signals.stopMovement.connect(self.stop_move)
+
+        self._thread_pool.start(self._pygame_joystick_runner)
+
+    @Slot()
+    def disconnect_controller(self):
+        if isinstance(self.mg, MotionGroup) and self.mg.is_moving:
+            self.stop_move()
+
+        if self._pygame_joystick_runner is None:
+            return
+
+        self._pygame_joystick_runner.quit()
+        self._pygame_joystick_runner = None
+        self._thread_pool.waitForDone(200)
+        self._thread_pool.clear()
+
+        if isinstance(self.mg, MotionGroup) and self.mg.is_moving:
+            self.stop_move()
+
+    def stop_move(self, axis=None, soft=False):
+        self.logger.debug("Stopping move.")
+
+        if axis is None:
+            self.mg.stop(soft=soft)
+            return
+
+        try:
+            self.mg.drive.send_command("stop", soft, axis=axis)
+        except Exception:  # noqa
+            self.mg.stop()
+
+    def zero_drive(self):
+        self.mg.set_zero()
+
+    @Slot()
+    def _drive_connection_lost(self):
+        super()._drive_connection_lost()
+        self.disconnect_controller()
+
+    @Slot(bool)
+    def _update_connect_led(self, value):
+        self.connected_led.setChecked(value)
+
+    @Slot(int, float)
+    def _handle_axis_move(self, jaxis, value):
+        if jaxis not in (1, 3):
+            # moved joystick axis is not utilized
+            return
+        elif jaxis == 1:
+            axis_id = 1
+        else:  # jaxis == 3:
+            axis_id = 0
+
+        ax = self.mg.drive.axes[axis_id]
+
+        if np.absolute(value) < 0.5:
+            self.stop_move(axis=axis_id, soft=True)
+        elif ax.is_moving:
+            pass
+        else:
+            try:
+                proceed = self.mspace_warning_dialog.exec()
+            except AttributeError:
+                proceed = False
+
+            if not proceed:
+                return
+
+            # pygame up-down axes are inverted
+            sign = 1 if value <= 0 else -1
+            sign = self.mspace_drive_polarity[axis_id] * sign
+            direction = "forward" if sign > 0 else "backward"
+
+            self.mg.drive.send_command("continuous_jog", direction, axis=axis_id)
+
+    @Slot(int)
+    def _handle_button_press(self, button):
+        if button in (0, 1):
+            self.stop_move()
+        elif button == 3:
+            self.zero_drive()
+
+    @Slot(int, int)
+    def _handle_hat_press(self, hat_id, direction):
+        if direction == 0:
+            # hat (dpad) button returned to unpressed state
+            # do nothing
+            return
+
+        try:
+            proceed = self.mspace_warning_dialog.exec()
+        except AttributeError:
+            proceed = False
+
+        if not proceed:
+            return
+
+        acw = self._axis_control_widgets[hat_id]
+        if direction > 0:
+            acw.jog_forward()
+        else:
+            acw.jog_backward()
+
+    def closeEvent(self, event):
+        self.disconnect_controller()
+        self._thread_pool.deleteLater()
+        super().closeEvent(event)
